Refactored packages accordingly.
