+++ app/src/main/java/ar/edu/itba/pam/travelapp/di/main/TripContainer.java
@@ -1,11 +1,9 @@
-package ar.edu.itba.pam.travelapp.model.di;
+package ar.edu.itba.pam.travelapp.di.main;
 
 import android.content.Context;
 
 import ar.edu.itba.pam.travelapp.landing.storage.FtuStorage;
-import ar.edu.itba.pam.travelapp.landing.storage.NightModeStorage;
-import ar.edu.itba.pam.travelapp.main.MainPresenter;
-import ar.edu.itba.pam.travelapp.main.MainView;
+import ar.edu.itba.pam.travelapp.main.storage.NightModeStorage;
 import ar.edu.itba.pam.travelapp.model.trip.TripRepository;
 import ar.edu.itba.pam.travelapp.utils.SchedulerProvider;
