+++ app/src/main/java/ar/edu/itba/pam/travelapp/di/main/TripModule.java
@@ -1,11 +1,11 @@
-package ar.edu.itba.pam.travelapp.model.di;
+package ar.edu.itba.pam.travelapp.di.main;
 
 import android.content.Context;
 
 import ar.edu.itba.pam.travelapp.landing.storage.FtuStorage;
-import ar.edu.itba.pam.travelapp.landing.storage.NightModeStorage;
+import ar.edu.itba.pam.travelapp.main.storage.NightModeStorage;
 import ar.edu.itba.pam.travelapp.landing.storage.SharedPreferencesFTUStorage;
-import ar.edu.itba.pam.travelapp.landing.storage.SharedPreferencesNightModeStorage;
+import ar.edu.itba.pam.travelapp.main.storage.SharedPreferencesNightModeStorage;
 import ar.edu.itba.pam.travelapp.model.AppDatabase;
 import ar.edu.itba.pam.travelapp.model.trip.TripDao;
 import ar.edu.itba.pam.travelapp.model.trip.TripMapper;
