+++ app/src/main/java/ar/edu/itba/pam/travelapp/main/storage/NightModeStorage.java
@@ -1,4 +1,4 @@
-package ar.edu.itba.pam.travelapp.landing.storage;
+package ar.edu.itba.pam.travelapp.main.storage;
 
 public interface NightModeStorage {
     boolean isActive();
